Rename class and move constants
