+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/PoetExtension.java
@@ -27,13 +27,6 @@ import software.amazon.awssdk.codegen.model.intermediate.ShapeType;
  */
 public class PoetExtension {
 
-    public static final String ASYNC_STREAMING_INPUT_PARAM = "requestBody";
-    public static final String ASYNC_STREAMING_OUTPUT_PARAM = "asyncResponseTransformer";
-    public static final String SYNC_STREAMING_INPUT_PARAM = "requestBody";
-    public static final String SYNC_STREAMING_OUTPUT_PARAM = "responseTransformer";
-    public static final String EVENT_PUBLISHER_PARAM_NAME = "requestStream";
-    public static final String EVENT_RESPONSE_HANDLER_PARAM_NAME = "asyncResponseHandler";
-
     private final IntermediateModel model;
 
     public PoetExtension(IntermediateModel model) {
