+++ services/s3/src/main/resources/codegen-resources/customization.config
@@ -172,5 +172,5 @@
       "clientConfiguration"
     ]
   },
-  "asyncClientDecoratorClass": true
+  "delegateAsyncClientClass": true
 }
